Deprecate `command` in favor of `send command`
